feat: add gift-card purchase on test endpoint.

+++ services/provider/giftcards/reloadly_models/gift_card_purchase_request.go
@@ -3,8 +3,8 @@ package reloadlymodels
 type GiftCardPurchaseRequest struct {
 	ProductID             int64                 `json:"productId"`
 	CountryCode           string                `json:"countryCode"`
-	Quantity              int64                 `json:"quantity"`
-	UnitPrice             int64                 `json:"unitPrice"`
+	Quantity              float64               `json:"quantity"`
+	UnitPrice             float64               `json:"unitPrice"`
 	CustomIdentifier      string                `json:"customIdentifier"`
 	SenderName            string                `json:"senderName"`
 	RecipientEmail        string                `json:"recipientEmail"`
